requirements.txt
inky[rpi]
Pillow
requests